staging: Move index update helper
Staging operations mostly build target content buffers for index and
working copy updates. update_index_with_blob_buffer is different: it writes
a blob, updates the Git index, and records journal details.

Keeping that Git mutation helper in staging.operations made the module mix
content construction with index side effects. That also forced callers that
only update the index to import the broad operations module.

This commit addresses that by moving update_index_with_blob_buffer into
staging.index_update and updating production callers that mutate the index
to import the new focused module.

Follow-up test commits will move the staging tests and add an architecture
guard for the split.
