+++ src/git_stage_batch/commands/batch_source/text_file_actions.py
@@ -11,7 +11,7 @@ from ...core.buffer import (
 )
 from ...core.text_lifecycle import TextFileChangeType, normalized_text_change_type
 from ...data.file_modes import apply_git_file_mode
-from ...staging.operations import update_index_with_blob_buffer
+from ...staging.index_update import update_index_with_blob_buffer
 from ...utils.git_index import git_update_index
 from ...utils.git_repository import get_git_repository_root_path
 from ...utils.git_object_io import create_git_blob
